refactor(admin): split qa section modules
Why:
- keep the QA route and content imports stable while separating filters, progress, and problem rows
- reduce the mixed qa-sections file into focused modules behind the existing barrel

Verify:
- pnpm exec prettier --check apps/admin/src/app/quantercise/qa/page.tsx apps/admin/src/app/quantercise/qa/qa-content.tsx apps/admin/src/app/quantercise/qa/qa-sections.tsx apps/admin/src/app/quantercise/qa/qa-filters.tsx apps/admin/src/app/quantercise/qa/qa-problems.tsx apps/admin/src/app/quantercise/qa/qa-progress.tsx
- git diff --check -- apps/admin/src/app/quantercise/qa/page.tsx apps/admin/src/app/quantercise/qa/qa-content.tsx apps/admin/src/app/quantercise/qa/qa-sections.tsx apps/admin/src/app/quantercise/qa/qa-filters.tsx apps/admin/src/app/quantercise/qa/qa-problems.tsx apps/admin/src/app/quantercise/qa/qa-progress.tsx
- pnpm --filter @anipotts/admin typecheck
- pnpm --filter @anipotts/admin lint
- pnpm --filter @anipotts/admin build
- pnpm validate
